Create number guessing game with python
This repository features a basic number guessing game implemented in Python. In this game, you will first enter a range of numbers. The program will then randomly select a number within that range. Your goal is to guess the number chosen by the program within a limited number of tries.